Removed built in HTTP router from victor framework.
It is still possible for chat/store adapters to create their own HTTP
routers/listeners and with the new optional config structs one could pass
them a port or an already created http object. By removing this from the
framework it is impossible to accidentally start it and safer to run
multiple bots in one process and one one machine.